+++ docs/changelog.md
@@ -10,9 +10,9 @@ nav_order: 2
 
 ### Connection
 
-- Tunneling: Refactor reconnection logic. Immediate first reconnection attempt, prevent reconnect task leak.
-- Tunneling: Mitigate dropping frames while reconnecting.
-- Tunneling UDP: Schedule disconnect when receiving frames with invalid sequence numbers.
+- Tunnelling: Refactor reconnection logic. Immediate first reconnection attempt, prevent reconnect task leak.
+- Tunnelling: Mitigate dropping frames while reconnecting.
+- Tunnelling UDP: Schedule disconnect when receiving frames with invalid sequence numbers.
 
 # 3.10.1 Fix cover auto-stop 2025-11-09